requirements.txt
langchain
python-dotenv
ipykernel
langchain-huggingface
langchain-community
langchain-chroma
chromadb
langchain-core
streamlit